Add files via upload
Here were three ways to do the landslide data inventory in this research, which are: 1) field survey in 2020, 2) landslide events records from BPBD Kabupaten Mojokerto data collection and 3) 2011—2020 Google Earth Pro imagery interpretation. 
The data collection resulted in 383 landslide events shown in This Figure.

The use of machine learning to model the landslide probability can be considered as a typical binary classification. As a consequence, the exact same number of non-landslide spots are also needed. Landslide events are symbolized as “1” and non-landslide events are symbolized as “0”.